Rename 'components/TimeStackedBarChart.vue' to 'components/index/CardsReference/TestedNumber/Chart.vue'
